Issue #2583, #2600 - move Factory class out of Constraint annotation
